fix: address code review issues for JWT and crypto utilities
- Change Claims timestamp fields from usize to i64 for consistency
- Remove jti field generation from access tokens (use empty string)
- Update encrypt_api_key to return Result instead of panicking
- Update tests to handle encrypt_api_key's new Result return type

All tests pass, no clippy warnings.

+++ src-server/src/models/auth.rs
@@ -4,8 +4,8 @@ use serde::{Deserialize, Serialize};
 pub struct Claims {
     pub sub: String,      // user_id
     pub username: String, // username
-    pub exp: usize,       // expiry time
-    pub iat: usize,       // issued at
+    pub exp: i64,         // expiry time
+    pub iat: i64,         // issued at
     pub jti: String,      // JWT ID (for refresh token)
 }
